added AddProduct Second Step Design

+++ core/src/main/java/com/floriaapp/core/domain/model/provider_/order_details/User.kt
@@ -25,5 +25,5 @@ data class User(
     @SerializedName("verified")
     val verified: Boolean,
     @SerializedName("wallet_balance")
-    val walletBalance: Int
+    val walletBalance: Double
 )
